third_party/WebKit/LayoutTests/http/tests/devtools/elements/edit/shadow-dom-modify-chardata-expected.txt

Tests that elements panel updates shadow dom tree structure upon typing.


Running: testDumpInitial
========= Original ========
- <div id="container">
    - <input type="text" id="input1">
        - #shadow-root (user-agent)
            - <div id="text-field-container" pseudo="-webkit-textfield-decoration-container">
                - <div id="editing-view-port">
                      <div></div>
                  </div>
              </div>
      </input>
  </div>

Running: testAppend
======== Type text =========
- <div id="container">
    - <input type="text" id="input1">
        - #shadow-root (user-agent)
            - <div id="text-field-container" pseudo="-webkit-textfield-decoration-container">
                - <div id="editing-view-port">
                      <div>Bar</div>
                  </div>
              </div>
      </input>
  </div>
